+++ exp/insert_test.go
@@ -3,7 +3,6 @@ package exp
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -35,7 +34,6 @@ type insertExpressionTestSuite struct {
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withDifferentRecordTypes() {
-	t := iets.T()
 	type testRecord struct {
 		C string `db:"c"`
 	}
@@ -46,98 +44,156 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withDifferentReco
 		testRecord{C: "v1"},
 		Record{"c": "v2"},
 	)
-	assert.EqualError(t, err, "goqu: rows must be all the same type expected exp.testRecord got exp.Record")
+	iets.EqualError(err, "goqu: rows must be all the same type expected exp.testRecord got exp.Record")
 	_, err = NewInsertExpression(
 		testRecord{C: "v1"},
 		testRecord2{C: "v2"},
 	)
-	assert.EqualError(t, err, "goqu: rows must be all the same type expected exp.testRecord got exp.testRecord2")
+	iets.EqualError(err, "goqu: rows must be all the same type expected exp.testRecord got exp.testRecord2")
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withInvalidValue() {
-	t := iets.T()
 	_, err := NewInsertExpression(true)
-	assert.EqualError(t, err, "goqu: unsupported insert must be map, goqu.Record, or struct type got: bool")
+	iets.EqualError(err, "goqu: unsupported insert must be map, goqu.Record, or struct type got: bool")
+}
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withDifferentTypes() {
+	_, err := NewInsertExpression(Record{"a": "a1"}, true)
+	iets.EqualError(err, "goqu: rows must be all the same type expected exp.Record got bool")
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withNoValues() {
-	t := iets.T()
 	ie, err := NewInsertExpression()
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert)
-	assert.Equal(t, eie, ie)
-	assert.True(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.True(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_Vals() {
+	ie, err := NewInsertExpression()
+	iets.NoError(err)
+	vals := [][]interface{}{
+		{"a", "b"},
+	}
+	ie = ie.SetCols(NewColumnListExpression("a", "b")).SetVals(vals)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+	iets.Equal(vals, ie.Vals())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_Cols() {
+	ie, err := NewInsertExpression()
+	iets.NoError(err)
+	vals := [][]interface{}{
+		{"a", "b"},
+	}
+	ce := NewColumnListExpression("a", "b")
+	ie = ie.SetCols(ce).SetVals(vals)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+	iets.Equal(vals, ie.Vals())
+	iets.Equal(ce, ie.Cols())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_From() {
+	ae := newTestAppendableExpression("select * from test", []interface{}{})
+	ie, err := NewInsertExpression(ae)
+	iets.NoError(err)
+	iets.False(ie.IsEmpty())
+	iets.True(ie.IsInsertFrom())
+	iets.Equal(ae, ie.From())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_appendableExpression() {
-	t := iets.T()
-
 	ae := newTestAppendableExpression("test ae", nil)
 
 	ie, err := NewInsertExpression(ae)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := &insert{from: ae}
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.True(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.True(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withRecords() {
-	t := iets.T()
 	ie, err := NewInsertExpression(Record{"c": "a"}, Record{"c": "b"})
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("c")).
 		SetVals([][]interface{}{{"a"}, {"b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withRecordsSlice() {
+	ie, err := NewInsertExpression([]Record{{"c": "a"}, {"c": "b"}})
+	iets.NoError(err)
+	eie := new(insert).
+		SetCols(NewColumnListExpression("c")).
+		SetVals([][]interface{}{{"a"}, {"b"}})
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withRecordOfDifferentLength() {
-	t := iets.T()
 	_, err := NewInsertExpression(Record{"c": "a"}, Record{"c": "b", "c2": "d"})
-	assert.EqualError(t, err, "goqu: rows with different value length expected 1 got 2")
+	iets.EqualError(err, "goqu: rows with different value length expected 1 got 2")
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withRecordWithDifferentkeys() {
-	t := iets.T()
 	_, err := NewInsertExpression(Record{"c1": "a"}, Record{"c2": "b"})
-	assert.EqualError(t, err, `goqu: rows with different keys expected ["c1"] got ["c2"]`)
+	iets.EqualError(err, `goqu: rows with different keys expected ["c1"] got ["c2"]`)
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withMap() {
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		map[string]interface{}{"c": "a"},
 		map[string]interface{}{"c": "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("c")).
 		SetVals([][]interface{}{{"a"}, {"b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructs() {
 	type testRecord struct {
 		C string `db:"c"`
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		testRecord{C: "a"},
 		testRecord{C: "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("c")).
 		SetVals([][]interface{}{{"a"}, {"b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructSlice() {
+	type testRecord struct {
+		C string `db:"c"`
+	}
+	ie, err := NewInsertExpression([]testRecord{
+		{C: "a"},
+		{C: "b"},
+	})
+	iets.NoError(err)
+	eie := new(insert).
+		SetCols(NewColumnListExpression("c")).
+		SetVals([][]interface{}{{"a"}, {"b"}})
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithoutTags() {
@@ -146,18 +202,17 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithou
 		FieldB bool
 		FieldC string
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		testRecord{FieldA: 1, FieldB: true, FieldC: "a"},
 		testRecord{FieldA: 2, FieldB: false, FieldC: "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("fielda", "fieldb", "fieldc")).
 		SetVals([][]interface{}{{int64(1), true, "a"}, {int64(2), false, "b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsIgnoredDbTag() {
@@ -166,18 +221,17 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsIgnore
 		FieldB bool
 		FieldC string
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		testRecord{FieldA: 1, FieldB: true, FieldC: "a"},
 		testRecord{FieldA: 2, FieldB: false, FieldC: "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("fieldb", "fieldc")).
 		SetVals([][]interface{}{{true, "a"}, {false, "b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithGoquSkipInsert() {
@@ -186,36 +240,34 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithGo
 		FieldB bool   `goqu:"skipupdate"`
 		FieldC string `goqu:"skipinsert"`
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		testRecord{FieldA: 1, FieldB: true, FieldC: "a"},
 		testRecord{FieldA: 2, FieldB: false, FieldC: "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("fielda", "fieldb")).
 		SetVals([][]interface{}{{int64(1), true}, {int64(2), false}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructPointers() {
 	type testRecord struct {
 		C string `db:"c"`
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		&testRecord{C: "a"},
 		&testRecord{C: "b"},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("c")).
 		SetVals([][]interface{}{{"a"}, {"b"}})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEmbeddedStructs() {
@@ -228,14 +280,13 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEm
 		Address string `db:"address"`
 		Name    string `db:"name"`
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		item{Address: "111 Test Addr", Name: "Test1", Phone: Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "211 Test Addr", Name: "Test2", Phone: Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "311 Test Addr", Name: "Test3", Phone: Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "411 Test Addr", Name: "Test4", Phone: Phone{Home: "123123", Primary: "456456"}},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("address", "home_phone", "name", "primary_phone")).
 		SetVals([][]interface{}{
@@ -244,9 +295,9 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEm
 			{"311 Test Addr", "123123", "Test3", "456456"},
 			{"411 Test Addr", "123123", "Test4", "456456"},
 		})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
 }
 
 func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEmbeddedStructPointers() {
@@ -259,14 +310,13 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEm
 		Address string `db:"address"`
 		Name    string `db:"name"`
 	}
-	t := iets.T()
 	ie, err := NewInsertExpression(
 		item{Address: "111 Test Addr", Name: "Test1", Phone: &Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "211 Test Addr", Name: "Test2", Phone: &Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "311 Test Addr", Name: "Test3", Phone: &Phone{Home: "123123", Primary: "456456"}},
 		item{Address: "411 Test Addr", Name: "Test4", Phone: &Phone{Home: "123123", Primary: "456456"}},
 	)
-	assert.NoError(t, err)
+	iets.NoError(err)
 	eie := new(insert).
 		SetCols(NewColumnListExpression("address", "home_phone", "name", "primary_phone")).
 		SetVals([][]interface{}{
@@ -275,9 +325,81 @@ func (iets *insertExpressionTestSuite) TestNewInsertExpression_withStructsWithEm
 			{"311 Test Addr", "123123", "Test3", "456456"},
 			{"411 Test Addr", "123123", "Test4", "456456"},
 		})
-	assert.Equal(t, eie, ie)
-	assert.False(t, ie.IsEmpty())
-	assert.False(t, ie.IsInsertFrom())
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withNilEmbeddedStructPointers() {
+	type Phone struct {
+		Primary string `db:"primary_phone"`
+		Home    string `db:"home_phone"`
+	}
+	type item struct {
+		*Phone
+		Address string `db:"address"`
+		Name    string `db:"name"`
+	}
+	ie, err := NewInsertExpression(
+		item{Address: "111 Test Addr", Name: "Test1"},
+		item{Address: "211 Test Addr", Name: "Test2"},
+		item{Address: "311 Test Addr", Name: "Test3"},
+		item{Address: "411 Test Addr", Name: "Test4"},
+	)
+	iets.NoError(err)
+	eie := new(insert).
+		SetCols(NewColumnListExpression("address", "name")).
+		SetVals([][]interface{}{
+			{"111 Test Addr", "Test1"},
+			{"211 Test Addr", "Test2"},
+			{"311 Test Addr", "Test3"},
+			{"411 Test Addr", "Test4"},
+		})
+	iets.Equal(eie, ie)
+	iets.False(ie.IsEmpty())
+	iets.False(ie.IsInsertFrom())
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withDifferentStructTypes() {
+	type Phone struct {
+		Primary string `db:"primary_phone"`
+		Home    string `db:"home_phone"`
+	}
+	type item struct {
+		*Phone
+		Address string `db:"address"`
+		Name    string `db:"name"`
+	}
+	_, err := NewInsertExpression(
+		item{Address: "111 Test Addr", Name: "Test1"},
+		Phone{Home: "123123", Primary: "456456"},
+		item{Address: "311 Test Addr", Name: "Test3"},
+		Phone{Home: "123123", Primary: "456456"},
+	)
+	iets.EqualError(err, "goqu: rows must be all the same type expected exp.item got exp.Phone")
+}
+
+func (iets *insertExpressionTestSuite) TestNewInsertExpression_withDifferentColumnLengths() {
+	type Phone struct {
+		Primary string `db:"primary_phone"`
+		Home    string `db:"home_phone"`
+	}
+	type Phone2 struct {
+		Primary string `db:"primary_phone2"`
+		Home    string `db:"home_phone2"`
+	}
+	type item struct {
+		*Phone
+		*Phone2
+		Address string `db:"address"`
+		Name    string `db:"name"`
+	}
+	_, err := NewInsertExpression(
+		item{Address: "111 Test Addr", Name: "Test1", Phone2: &Phone2{Home: "123123", Primary: "456456"}},
+		item{Address: "311 Test Addr", Name: "Test3", Phone: &Phone{Home: "123123", Primary: "456456"}},
+	)
+	iets.EqualError(err, `goqu: rows with different keys expected `+
+		`["address","home_phone2","name","primary_phone2"] got ["address","home_phone","name","primary_phone"]`)
 }
 
 func TestInsertExpressionSuite(t *testing.T) {